refactor(crane-context): split constants.ts by domain (plan §V)
Pure data move — no runtime changes. Splits 506-LOC monolith into
8 per-domain files under `workers/crane-context/src/constants/`:

- sessions.ts       (~10 LOC) — STALE_AFTER_MINUTES, HEARTBEAT_*, SESSION_STATUSES, END_REASONS
- notifications.ts  (~90 LOC) — NOTIFICATION_*, GREEN_CONCLUSIONS, GREEN_DEPLOYMENT_TYPES, SCHEDULE_LIKE_EVENTS, PROTECTED_BRANCHES, isProtectedBranch, VERCEL_PROJECT_TO_VENTURE
- ventures.ts       (~45 LOC) — VENTURE_CONFIG, VENTURES, Venture
- verify.ts         (~40 LOC) — all VERIFY_* constants and derived types
- notes.ts          (~20 LOC) — RECOMMENDED_TAGS, KNOWLEDGE_BASE_TAGS, MAX_NOTE_CONTENT_SIZE
- docs.ts           (~90 LOC) — DEFAULT_DOC_REQUIREMENTS
- http.ts           (~15 LOC) — HTTP_STATUS
- limits.ts         (~25 LOC) — payload/body size limits, ID_PREFIXES, pagination defaults
- index.ts          (8 LOC)   — barrel re-export of all domain files

src/constants.ts is now a single-line barrel: export * from './constants/index.js'

Barrel re-export preserves all 24 existing import sites — no importer changes needed.

Mechanical gate: constants.ts drops out of >500 LOC list. No new files >500 LOC.
