+++ workers/crane-context/src/constants/docs.ts
@@ -1,0 +1,94 @@
+export const DEFAULT_DOC_REQUIREMENTS = [
+  {
+    doc_name_pattern: '{venture}-project-instructions.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: null,
+    auto_generate: true,
+    generation_sources: '["claude_md","readme","package_json","docs_process","ventures_json"]',
+    description: 'Project instructions - product vision, tech stack, principles, constraints.',
+    staleness_days: 30,
+  },
+  {
+    doc_name_pattern: '{venture}-api.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: 'has_api',
+    auto_generate: true,
+    generation_sources: '["route_files","openapi","tests"]',
+    description: 'API reference - endpoints, auth, request/response shapes.',
+    staleness_days: 60,
+  },
+  {
+    doc_name_pattern: '{venture}-schema.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: 'has_database',
+    auto_generate: true,
+    generation_sources: '["migrations","schema_files","wrangler_toml"]',
+    description: 'Database schema - tables, columns, relationships.',
+    staleness_days: 60,
+  },
+  // Portfolio docs (hub-sourced, synced from docs/ventures/{code}/)
+  {
+    doc_name_pattern: 'product-overview.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Product overview - what it is, target market, value prop, tech stack.',
+    staleness_days: 90,
+  },
+  {
+    doc_name_pattern: 'roadmap.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Product roadmap - current milestone, planned work, recent completions.',
+    staleness_days: 30,
+  },
+  {
+    doc_name_pattern: 'metrics.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Product metrics - KPIs, stage-appropriate measurements, health signals.',
+    staleness_days: 60,
+  },
+  {
+    doc_name_pattern: 'design-spec.md',
+    scope_type: 'all_ventures',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Design spec - tokens, colors, typography, component patterns, brand voice.',
+    staleness_days: 90,
+  },
+  // Global enterprise docs
+  {
+    doc_name_pattern: 'company-overview.md',
+    scope_type: 'global',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Company structure, mission, entity overview.',
+    staleness_days: 180,
+  },
+  {
+    doc_name_pattern: 'strategic-planning.md',
+    scope_type: 'global',
+    required: true,
+    condition: null,
+    auto_generate: false,
+    generation_sources: '[]',
+    description: 'Capital allocation principles, evaluation framework.',
+    staleness_days: 90,
+  },
+] as const
