+++ workers/crane-context/src/constants/notes.ts
@@ -1,0 +1,21 @@
+export const RECOMMENDED_TAGS = [
+  'executive-summary',
+  'prd',
+  'design',
+  'strategy',
+  'methodology',
+  'market-research',
+  'bio',
+  'marketing',
+  'governance',
+] as const
+
+export const KNOWLEDGE_BASE_TAGS = [
+  'prd',
+  'design',
+  'strategy',
+  'methodology',
+  'market-research',
+] as const
+
+export const MAX_NOTE_CONTENT_SIZE = 500 * 1024
